(MAINT) Fix tests for puppetserver pipelines
Currently the puppetserver CI jobs which run the puppet agent tests are failing

When puppetserver runs the puppet tests it uses a beaker config which creates
multiple machines with agent roles. This causes the 'agent' beaker variable to
be a list of hashes instead of a hash. Tests which access the agent array as if
it's a hash throw errors

This commit modifies the offending tests to work under both conditions